+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to Kawaii Cocoa Grader will be documented here
 
 Changelog before 0.2.2 will not be noted here
 
+## [0.2.6] - 2021-01-30
+
+- Upgrade to Cocoa Discord Utils 1.0.0-rc.1 and use its new feature
+
+- Embed Submission Result ✨✨
+
 ## [0.2.5] - 2021-01-29
 
 - Upgrade to Cocoa Discord Utils 1.0.0-pre.10 and use its new feature